apps/api/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
greenlet>=3.0.0

# Redis
redis>=5.0.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0

# Quantum computing
qiskit>=1.0.0
qiskit-aer>=0.13.0
qiskit-ibm-runtime>=0.20.0

# Validation and serialization
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Storage
boto3>=1.34.0

# Utilities
python-dotenv>=1.0.0
orjson>=3.9.10

# Development
pytest>=7.4.4
pytest-asyncio>=0.23.3
black>=24.1.0
ruff>=0.1.14
